requirements.txt
python-telegram-bot==21.4
feedparser==6.0.11
trafilatura==1.7.0
sumy==0.11.0
python-dotenv==1.0.1
requests==2.32.3


